clean up per-game rules and scoring guide
the scoring footnote showed klondike's 1-card/3-card time bonus text under
every variant. move all guide copy into GameGuide keyed by variant, state
each game's actual bonus, add hasSelectableDrawMode so only-klondike-has-
draw-modes lives in one place, rename the view to GameGuideView, fill
klondike's missing rules, and add tests that hold the copy to the engine's
behavior.
